+++ routes/web.php
@@ -72,11 +72,11 @@ Route::get('/archivoSem/{dias}', function ($dias) {
 dd(CronFunciones::generarArchivoSem($dias));
 });
 
-Route::get('/sarasa', function () {
-        //CronFunciones::diario();
-        CronFunciones::logError(['clase' => 'routes/web.php', 'metodo' => 'sarasa', 'error' => 'Funciona OK.']);
+/* Route::get('/sarasa', function () {
+        CronFunciones::actualizarIssuesVencidos ();
+        //CronFunciones::logError(['clase' => 'routes/web.php', 'metodo' => 'sarasa', 'error' => 'Funciona OK.']);
         //CronFunciones::enviarErrorsMail();
-});
+}); */
 
 /* Route::view('/welcome', 'index'); */
 /* Route::get('/test/', function () {